Refactoring Remove TabBarCreator Create RootCoordinator Add spinner in ProfileVC
